mold: fix `mold -run` on older systems
`mold -run` is a mold feature that intercepts exec(3) functions to
replace `ld` invocations with `mold`. It is supported by
`mold-wrapper.so`, a shared object file designed to be used with
`LD_PRELOAD`. Since it can be loaded with arbitrary binaries outside of
Homebrew, we need to make sure its symbols are compatible with the
system linker. Otherwise, on an older system like Ubuntu 20.04, it can
fail to load with the following error:

    $ mold -run echo hello
    echo: /lib/x86_64-linux-gnu/libc.so.6: version `GLIBC_2.34' not found (required by /home/linuxbrew/.linuxbrew/Cellar/mold/2.30.0/bin/../lib/mold/mold-wrapper.so)

The error is a result of a bump of the `dlsym` symbol's version in glibc
2.34. We can force `mold-wrapper.so` to use an older version of the
symbol by adding a `.symver` directive to its source code.

In addition, as `dlsym` has been integrated into libc in glibc 2.34,
libdl is not linked anymore. To allow it to be used on older systems, we
need to make sure libdl is added as a dependency. This does not affect
systems with glibc 2.34 or newer, because libdl is kept as an empty
library.

See also:
https://github.com/bminor/glibc/commit/77f876c0e3ac08a98daa60fbad44061d4e4c3d14
https://github.com/rui314/mold/blob/v2.30.0/docs/mold.md#mold-specific-options
https://sourceware.org/binutils/docs/ld/VERSION.html